removed PKIS2 since no longer necessary - everything pre-processed in mkt.databases.datasets; fintune now contains KinaseSplit and CrossValidation classes
